backend/twitter/serializers.py
from .models import User, Status
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'name', 'screen_name']


class StatusSerializer(serializers.HyperlinkedModelSerializer):
    user = UserSerializer()

    class Meta:
        model = Status
        fields = ['user','id', 'created_at','content']
