backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8.6,<2.9
djangorestframework
django-cors-headers
tweepy
pandas
snscrape
spacy